1114|소셜로그인

+++ 05_Django/04_django_form/articles/views.py
@@ -1,13 +1,14 @@
+import hashlib
+from itertools import chain
 from IPython import embed
+from django.contrib.auth import get_user_model
 from django.contrib.auth.decorators import login_required
 from django.views.decorators.http import require_POST
 from django.shortcuts import render, redirect, get_object_or_404
-from .models import Article, Comment
-
+from .models import Article, Comment, Hashtag
 from .forms import ArticleForm, CommentForm
 # Create your views here.
 def index(request):
-    # embed()
     articles = Article.objects.all()
     context = {'articles': articles,}
     return render(request, 'articles/index.html', context)
@@ -18,7 +19,17 @@ def create(request):
         # embed()
         form = ArticleForm(request.POST)
         if form.is_valid():
-            article = form.save()
+            article = form.save(commit=False)
+            article.user = request.user
+            article.save()
+            # hashtag
+            # 게시글 내용을 split 리스트로 만듦
+            for word in article.content.split():
+              # word가 '#'으로 시작할 경우 해시태그 등록
+              if word.startswith('#'):
+                hashtag, created = Hashtag.objects.get_or_create(content=word)
+                article.hashtags.add(hashtag)
+
         return redirect('articles:detail', article.pk)
     else:
         form = ArticleForm()
@@ -27,10 +38,13 @@ def create(request):
 
 def detail(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
+
+    person = get_object_or_404(get_user_model(), pk=article.user_id)
     comment_form = CommentForm()
     comments = article.comment_set.all()
     context = {
         'article': article,
+        'person': person,
         'comment_form': comment_form,
         'comments': comments,
     }
@@ -38,21 +52,37 @@ def detail(request, article_pk):
 
 @require_POST
 def delete(request, article_pk):
+    # 지금 사용자가 로그인 되어있는지?
     if request.user.is_authenticated:
+        # 삭제할 게시글 가져오기
         article = get_object_or_404(Article, pk=article_pk)
-        article.delete()
+        # 로그인한 사용자와 게시글 작성자 비교
+        if request.user == article.user:
+            article.delete()
+        else:
+            return redirect('articles:detail', article.pk)
     return redirect('articles:index')
 
 @login_required
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
-    if request.method == 'POST':
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():
-            article = form.save()
-            return redirect('articles:detail', article.pk)
+    if request.user == article.user:
+        if request.method == 'POST':
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid():
+                article = form.save()
+                # hashtag
+                article.hashtags.clear()
+                for word in article.content.split():
+                  if word.startswith('#'):
+                    hashtag, created = Hashtag.objects.get_or_create(content=word)
+                    article.hashtags.add(hashtag)
+
+                return redirect('articles:detail', article.pk)
+        else:
+            form = ArticleForm(instance=article)
     else:
-        form = ArticleForm(instance=article)
+        return redirect('articles:index')
     context = {
         'form': form,
         'article': article
@@ -61,18 +91,88 @@ def update(request, article_pk):
 
 @require_POST
 def comments_create(request, article_pk):
-    article = get_object_or_404(Article, pk=article_pk)
+    # article = get_object_or_404(Article, pk=article_pk)
     if request.user.is_authenticated:
         comment_form = CommentForm(request.POST)
         if comment_form.is_valid():
             comment = comment_form.save(commit=False)
-            comment.article = article
+            comment.user = request.user
+            comment.article_id = article_pk
             comment.save()
-    return redirect('articles:detail', article.pk)
+    return redirect('articles:detail', article_pk)
 
 @require_POST
 def comments_delete(request, article_pk, comment_pk):
+    # 1. 로그인 여부 확인
     if request.user.is_authenticated:
         comment = get_object_or_404(Comment, pk=comment_pk)
-        comment.delete()
+        # 2. 로그인한 사용자와 댓글 작성자가 같을 경우
+        if request.user == comment.user:
+            comment.delete()
     return redirect('articles:detail', article_pk)
+
+@login_required
+def like(request, article_pk):
+    article = get_object_or_404(Article, pk=article_pk)
+    user = request.user
+    if article.like_users.filter(pk=user.pk).exists():
+        article.like_users.remove(user)
+    else:
+        article.like_users.add(user)
+    return redirect('articles:index')
+
+@login_required
+def follow(request, article_pk, user_pk):
+    # 게시글 작성한 유저
+    person = get_object_or_404(get_user_model(), pk=user_pk)
+    # 지금 접속하고 있는 유저
+    user = request.user
+
+    # 게시글 작성 유저의 팔로워 목록에 접속 중인 유저가 있을 경우
+    # -> Unfollow
+    if person != user:
+      if user in person.followers.all():
+        person.followers.remove(user)
+      # 목록에 없으면 -> Follow
+      else:
+        person.followers.add(user)
+    # 게시글 상세정보로 redirect
+    return redirect('articles:detail', article_pk)
+
+# 내가 팔로우 하는 사람의 글 + 내가 작성한 글
+def list(request):
+  # 내가 팔로우하고 있는 사람들
+  followings = request.user.followings.all()
+  # 내가 팔로우하고 있는 사람들 + 나 -> 합치기
+  followings = chain(followings, [request.user])
+  # 위 명단 사람들 게시글 가져오기
+  articles = Article.objects.filter(user__in=followings).order_by('-pk').all()
+
+  comment_form = CommentForm()
+  context = {
+    'articles': articles,
+    'comment_form':comment_form,
+  }
+  return render(request, 'articles/article_list.html', context)
+
+# 모든 사람 글
+def explore(request):
+  articles = Article.objects.all()
+  comment_form = CommentForm()
+  context = {
+    'articles':articles,
+    'comment_form':comment_form,
+  }
+  return render(request, 'articles/article_list.html', context)
+
+# Hashtag 글 모아보기
+def hashtag(request, hash_pk):
+  # 해시태그 가져오기
+  hashtag = get_object_or_404(Hashtag, pk=hash_pk)
+  # 해당 해시태그를 참조하는 게시글들 가져오기
+  articles = hashtag.article_set.order_by('-pk')
+  content = {
+    'hashtag': hashtag,
+    'articles': articles,
+  }
+  return render(request, 'articles/hashtag.html',content)
